requirements-dev.txt
-e requirements.txt
black==23.9.1
coverage==7.3.1
isort==5.12.0
pylint==2.17.5
pylint-django==2.5.3
